pyproject.toml
[tool.poetry]
name = "sesame"
version = "0.1.0"
description = "slack bot for requesting IAM permissions"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.15.5"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
